Move Player widgets to their own file (ui/player/widgets.py)
- it simplifies imports and shortens files
